+++ selcukes-commons/src/main/java/io/github/selcukes/commons/helper/ServiceLoaderUtils.java
@@ -17,7 +17,7 @@
 package io.github.selcukes.commons.helper;
 
 import io.github.selcukes.commons.exception.ConfigurationException;
-import io.github.selcukes.databind.utils.Streams;
+import io.github.selcukes.databind.collections.Streams;
 import lombok.experimental.UtilityClass;
 
 import java.util.List;
