+++ selcukes-databind/src/main/java/io/github/selcukes/databind/collections/DataTable.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package io.github.selcukes.databind.utils;
+package io.github.selcukes.databind.collections;
 
 import lombok.Getter;
 import lombok.NonNull;
