+++ selcukes-databind/src/main/java/io/github/selcukes/databind/collections/Lists.java
@@ -14,8 +14,9 @@
  *  limitations under the License.
  */
 
-package io.github.selcukes.databind.utils;
+package io.github.selcukes.databind.collections;
 
+import io.github.selcukes.databind.utils.StringHelper;
 import lombok.NoArgsConstructor;
 import lombok.experimental.UtilityClass;
