+++ selcukes-databind/src/test/java/io/github/selcukes/databind/tests/DataTableTest.java
@@ -16,7 +16,7 @@
 
 package io.github.selcukes.databind.tests;
 
-import io.github.selcukes.databind.utils.DataTable;
+import io.github.selcukes.databind.collections.DataTable;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
