+++ selcukes-excel-runner/src/main/java/io/github/selcukes/excel/SingleExcelData.java
@@ -19,8 +19,8 @@ package io.github.selcukes.excel;
 import io.github.selcukes.commons.config.ConfigFactory;
 import io.github.selcukes.commons.exception.ExcelConfigException;
 import io.github.selcukes.databind.excel.ExcelMapper;
-import io.github.selcukes.databind.utils.Maps;
-import io.github.selcukes.databind.utils.Streams;
+import io.github.selcukes.databind.collections.Maps;
+import io.github.selcukes.databind.collections.Streams;
 import io.github.selcukes.databind.utils.StringHelper;
 import lombok.CustomLog;
 import lombok.experimental.UtilityClass;
